+++ cpp/open3d/t/geometry/TriangleMesh.h
@@ -29,8 +29,8 @@
 #include "open3d/core/Tensor.h"
 #include "open3d/core/TensorList.h"
 #include "open3d/geometry/TriangleMesh.h"
-#include "open3d/tgeometry/Geometry.h"
-#include "open3d/tgeometry/TensorListMap.h"
+#include "open3d/t/geometry/Geometry.h"
+#include "open3d/t/geometry/TensorListMap.h"
 
 namespace open3d {
 namespace tgeometry {
