+++ cpp/pybind/t/geometry/geometry.cpp
@@ -3,7 +3,7 @@
 // ----------------------------------------------------------------------------
 // The MIT License (MIT)
 //
-// Copyright (c) 2020 www.open3d.org
+// Copyright (c) 2018 www.open3d.org
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -24,38 +24,32 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/tgeometry/TensorListMap.h"
+#include "open3d/t/geometry/Geometry.h"
 
-#include "open3d/tgeometry/PointCloud.h"
 #include "pybind/docstring.h"
-#include "pybind/tgeometry/geometry.h"
+#include "pybind/t/geometry/geometry.h"
 
 namespace open3d {
 namespace tgeometry {
 
-void pybind_tensorlistmap(py::module& m) {
-    // Bind to the generic dictionary interface such that it works the same as a
-    // regular dictionay in Python, except that types are enforced. Supported
-    // functions include `__bool__`, `__iter__`, `items`, `__getitem__`,
-    // `__contains__`, `__delitem__`, `__len__` and map assignment.
-    auto tlm = py::bind_map<TensorListMap>(m, "TensorListMap",
-                                           "Map of TensorList by string.");
+void pybind_geometry_class(py::module& m) {
+    py::class_<Geometry, PyGeometry<Geometry>, std::unique_ptr<Geometry>>
+            geometry(m, "Geometry", "The base geometry class.");
 
-    // Constructors.
-    tlm.def(py::init<const std::string&>(), "primary_key"_a)
-            .def(py::init<const std::string&,
-                          const std::unordered_map<std::string,
-                                                   core::TensorList>&>(),
-                 "primary_key"_a, "map_keys_to_tensorlists"_a);
+    geometry.def("clear", &Geometry::Clear,
+                 "Clear all elements in the geometry.")
+            .def("is_empty", &Geometry::IsEmpty,
+                 "Returns ``True`` iff the geometry is empty.");
+    docstring::ClassMethodDocInject(m, "Geometry", "clear");
+    docstring::ClassMethodDocInject(m, "Geometry", "is_empty");
+}
 
-    // Member functions. Some C++ functions are ignored since the
-    // functionalities are already covered in the generic dictionary interface.
-    tlm.def("synchronized_push_back", &TensorListMap::SynchronizedPushBack,
-            "map_keys_to_tensors"_a)
-            .def("get_primary_key", &TensorListMap::GetPrimaryKey)
-            .def("is_size_synchronized", &TensorListMap::IsSizeSynchronized)
-            .def("assert_size_synchronized",
-                 &TensorListMap::AssertSizeSynchronized);
+void pybind_geometry(py::module& m) {
+    py::module m_submodule = m.def_submodule("tgeometry");
+
+    pybind_geometry_class(m_submodule);
+    pybind_tensorlistmap(m_submodule);
+    pybind_pointcloud(m_submodule);
 }
 
 }  // namespace tgeometry
