+++ cpp/tests/t/geometry/TriangleMesh.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/tgeometry/TriangleMesh.h"
+#include "open3d/t/geometry/TriangleMesh.h"
 
 #include "core/CoreTest.h"
 #include "open3d/core/TensorList.h"
